+++ src/main/java/io/supertokens/webserver/WebserverAPI.java
@@ -155,10 +155,10 @@ public abstract class WebserverAPI extends HttpServlet {
             if (this.versionNeeded(req)) {
                 String version = getVersionFromRequest(req);
                 assertThatVersionIsCompatible(version);
-                Logging.debug(main,
+                Logging.info(main,
                         "API called: " + this.getPath() + ". Method: " + req.getMethod() + ". Version: " + version);
             } else {
-                Logging.debug(main, "API called: " + this.getPath() + ". Method: " + req.getMethod());
+                Logging.info(main, "API called: " + this.getPath() + ". Method: " + req.getMethod());
             }
             super.service(req, resp);
         } catch (Exception e) {
@@ -180,7 +180,7 @@ public abstract class WebserverAPI extends HttpServlet {
                 sendTextResponse(500, "Internal Error", resp);
             }
         }
-        Logging.debug(main, "API ended: " + this.getPath() + ". Method: " + req.getMethod());
+        Logging.info(main, "API ended: " + this.getPath() + ". Method: " + req.getMethod());
     }
 
     protected String getRIDFromRequest(HttpServletRequest req) {